Element app views renamed
